src/requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
Django==3.0.5
django-oauth-toolkit==1.3.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.9
oauthlib==3.1.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
